error handling

+++ app/index.js
@@ -26,42 +26,59 @@ app.use(express.static(path.join(__dirname, '/../public')))
 
 
 app.get('/', async (req, res) => {
-	let input = fs.readFileSync(__dirname + '/../data/updated.csv')
-	var data = csvParseSync(input)
-	data = csvToJson(data)
-	data = data.map((d) => {
-		d.Quantity = parseFloat(d.Quantity)
-		d['Purchase Price'] = parseFloat(d['Purchase Price'])
-		if(d.enrich){
-			d.enrich = JSON.parse(d.enrich)
-		}
-		return d
-	})
+	let inputPath = __dirname + '/../data/updated.csv'
+	if(!fs.existsSync(inputPath)){
+		res.render('main', {
+		})
+	}
 
-	let calculated = calc(data)
-	let port = calculated.arr
-	let stats = calculated.stats
+	else{
+		let input = fs.readFileSync(inputPath)
+		var data = csvParseSync(input)
+		data = csvToJson(data)
+		data = data.map((d) => {
+			d.Quantity = parseFloat(d.Quantity)
+			d['Purchase Price'] = parseFloat(d['Purchase Price'])
+			if(d.enrich){
+				d.enrich = JSON.parse(d.enrich)
+			}
+			return d
+		})
 
-	res.render('main', {
-		data: data,
-		port: port,
-		stats: stats,
-		roundTo2: roundTo2
-	})
+		let calculated = calc(data)
+		let port = calculated.arr
+		let stats = calculated.stats
+
+		res.render('main', {
+			data: data,
+			port: port,
+			stats: stats,
+			roundTo2: roundTo2
+		})
+	}
+
 })
 
 
 app.get('/update', async (req, res) => {
-	let input = fs.readFileSync(__dirname + '/../data/data.csv')
-	var data = csvParseSync(input)
-	data = csvToJson(data)
-	data = await getPrices(data)
-	data.sort((a, b) => {
-		return a['Trade Date'] - b['Trade Date']
-	})
-	console.log(data)
-	fs.writeFileSync(__dirname + '/../data/updated.csv', stringify(data, {header: true}))
-	res.send('done. <a href="/">Go back</a>')
+	let inputPath = __dirname + '/../data/data.csv'
+	if(!fs.existsSync(inputPath)){
+		res.send(`Please create data/data.csv and data/initial_amount input files before updating.<br><a href="/">Go back</a>`)
+	}
+
+	else{
+		let input = fs.readFileSync(__dirname + '/../data/data.csv')
+		var data = csvParseSync(input)
+		data = csvToJson(data)
+		data = await getPrices(data)
+		data.sort((a, b) => {
+			return a['Trade Date'] - b['Trade Date']
+		})
+		console.log(data)
+		fs.writeFileSync(__dirname + '/../data/updated.csv', stringify(data, {header: true}))
+		res.send('done. <a href="/">Go back</a>')		
+	}
+
 })
 
 app.listen(3000, (err) => {
